Refactor UI and Video Functions
- Updated PAL_PlayerInfoBox to remove unused parameters.
- Commented out unnecessary VIDEO_UpdateScreen calls in various menu functions to improve performance.
- Adjusted key input handling to use FRAME_TIME for consistency.
- Modified VIDEO_UpdateScreen to accept a boolean for shake effect instead of an integer.
- Simplified VIDEO_SwitchScreen and VIDEO_FadeScreen functions by removing speed parameter and adjusting delay handling.
- Changed PAL_Rect structure to use unsigned short for better memory management.

+++ src/fight.h
@@ -21,7 +21,7 @@
 #ifndef FIGHT_H
 #define FIGHT_H
 
-int PAL_BattleSelectAutoTargetFrom(
+short PAL_BattleSelectAutoTargetFrom(
     int);
 
 int PAL_IsPlayerDying(
